Updated project and documentation.
