+++ docs/source/index.rst
@@ -1,8 +1,1 @@
-.. mdinclude:: ../../README.md
 .. include:: contents.rst
-
-Indices and tables
-==================
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
