--- docs/source/module/helpers.rst
@@ -1,36 +1,0 @@
-Helpers
-=======
-
-.. automodule:: httpsuite.helpers
-
-----
-
-Item
-****
-
-.. autoclass:: httpsuite.helpers.Item
-  :members:
-
-----
-
-Headers
-*******
-
-.. autoclass:: httpsuite.helpers.Headers
-  :members:
-
-----
-
-TwoWayFrozenDict
-****************
-
-.. autoclass:: httpsuite.helpers.TwoWayFrozenDict
-  :members:
-
-----
-
-FrozenSet
-*********
-
-.. autoclass:: httpsuite.helpers.FrozenSet
-  :members:
