+++ docs/source/writeups/Quickstart.rst
@@ -1,80 +1,91 @@
-**********
+##########
 Quickstart
-**********
+##########
 
+
+**********
 Installing
 **********
 
-To get started with ``httpsuite``, install the latest stable release from
-`PyPi <https://pypi.org/project/httpsuite/>`_:
+To get started with ``httpsuite``, install the latest stable release from `PyPi <https://pypi.org/project/httpsuite/>`_:
 
 .. code-block:: bash
 
   pip install httpsuite
 
-
-If you would like to use the latest non-stable release, you may install directly
-from Github:
-
-.. code-block:: bash
-
-  pip install git+https://github.com/synchronizing/httpsuite
-
+***************
 Getting Started
 ***************
 
-There are two main abstractions that ``httpsuite`` provides, ``Request`` and
-``Response``. Both provide an interface to interpret, modify, and compile
-raw HTTP messages.
+There are two principal abstractions to be aware of it in ``httpsuite``, ``Request`` and ``Response``.
 
 .. code-block:: python
 
   from httpsuite import Request, Response
 
-You may either chose to initialize them via their ``__init__`` methods:
+These classes represent an HTTP/1.x request and response message, and offer a high-level API to :term:`modify`, 
+:term:`compile`, and :term:`parse` them. The :class:`Request` and :class:`Response` can be initialized via two different
+methods.
+
+``__init__()``
+==============
+
+Create a :class:`Request` or :class:`Response` object with the given parameter.
 
 .. code-block:: python
+  
+  from httpsuite import Request, Response
 
   request = Request(
-      method="GET",
+      method="POST",
       target="/",
       protocol="HTTP/1.1",
-      headers=None,
-      body=None
+      headers={"User-Agent": "httpsuite", "Content-Length": 12},
+      body="Hello world."
   )
 
   response = Response(
       protocol="HTTP/1.1",
       status=200,
       status_msg="OK",
-      headers=None,
-      body=None
+      headers={"User-Agent": "httpsuite", "Content-Length": 8},
+      body="Hi back!"
   )
 
-Or, parse them from raw HTTP messages via their ``parse()`` method:
+``.parse()``
+============
+
+Or, parse from `bytes` to create a new :class:`Request` or :class:`Response` with the given details.
+
+``bytes``
+---------
+
+Most useful when using socket connections.
 
 .. code-block:: python
 
-  request = Request.parse(b'GET / HTTP/1.1\r\n\r\n')
-  response = Response.parse(b'HTTP/1.1 200 OK\r\n\r\n')
+  from httpsuite import Request, Response
 
-Once you have initialized a ``Request`` or ``Response`` message, you can modify
-and then compile.
+  req = Request.parse(b"GET / HTTP/1.1\r\nUser-Agent: httpsuite\r\nContent-Length: 12\r\n\r\nHello world")
+  resp = Response.parse(b"HTTP/1.1 200 OK\r\nUser-Agent: httpsuite\r\nContent-Length: 12\r\n\r\nHi back!")
 
-Modifying Messages
-******************
+******
+Modify
+******
 
-Modifying ``Request`` or ``Response`` objects is done as one would expect:
+The next probable step after initializing a :class:`Request` or :class:`Response` object is to :term:`modify` and 
+:term:`compile`. Object modification is done as one would expect.
 
 .. code-block:: python
 
   request.method = "POST"
-  response.status = b"300"
+  response.status = 300
   response.status_msg = b"Continue"
 
-Notice, however, that setting a ``Request`` or ``Response`` value is type-agnostic.
-In other words, it does not matter if you set a piece of data as a ``str``,
-``bytes``, or ``int``. Similarly, comparisons are type-agnostic:
+Notice that setting object properties is type-agnostic. Properties can be modified to either ``int``, ``str``, 
+or ``bytes`` objects. Internally, ``httpsuite`` automatically converts every property of a :class:`request` or
+:class:`response` into an :class:`Item`, which is a low-level interface to allow easy setting and comparissons on the
+fly. Similarly to setting properties, one can be assured of type-agnostic property comparissions.
 
 .. code-block:: python
 
@@ -82,15 +93,17 @@ In other words, it does not matter if you set a piece of data as a ``str``,
   request.status == "300"   # True
   request.status == b"300"  # True
 
-Compiling Message
-*****************
+*******
+Compile
+*******
 
-After modifying the message, the final step is to compile the message. Compiling
-is simply translating our object from Python to bytes (following specific rules),
-so that another computer on the internet can understand us. One does this by calling
-the property ``raw``:
+After modifying a message compilation allows the :class:`Request` and :class:`Response` objects to be compiled into 
+less maluable yet useful types. Those types being ``bytes`` or ``str``.
 
 .. code-block:: python
+
+  from httpsuite import Request, Response
+  import json
 
   body = json.dumps({"hello": "world"})
   request = Request(
@@ -106,17 +119,41 @@ the property ``raw``:
       body=body,
   )
 
+``.raw``
+=========
+
+Useful to use with sockets.
+
+.. code-block:: python
+
   print(request.raw)
 
 .. code-block:: python
 
   b'POST /post HTTP/1.1\r\nHost: httpbin.org\r\nConnection: close\r\nContent-Length: 18\r\nAccept: */*\r\n\r\n{"hello": "world"}'
 
-Example
-*******
+``__str__``
+===========
 
-A simple example of using ``httpsuite`` with sockets to send our message above:
+Pretty print of the object.
 
-.. literalinclude:: ../../../examples/example_sockets.py
-   :language: python
-   :linenos:
+.. code-block:: python
+
+  print(request)
+
+.. code-block::
+
+  → POST /post HTTP/1.1
+  → Host: httpbin.org
+  → Connection: close
+  → Content-Length: 18
+  → Accept: */*
+  → {"hello": "world"}
+
+
+****
+More 
+****
+
+If you finished this guide and want to continue learning more you can do so by reading the package's documentation found
+on the left menu.
