+++ httpsuite/__init__.py
@@ -1,7 +1,7 @@
-__version__ = "1.0.6"
-__IETF__ = "https://tools.ietf.org/html/"
-__ENCODE__ = "ascii"
+from toolbox.pkgutil import search_package
 
-from .core import *
-from .helpers import *
-from .RFC import *
+from .http import *
+from .interface import *
+from .rfc import *
+
+__plugins__ = search_package("httpsuite_", "startswith", True)
